LayoutTests/js/dom/global-constructors-attributes-dedicated-worker-expected.txt

Test to ensure that global constructors have the right attributes in the dedicated worker context

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: script-tests/global-constructors-attributes.js
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Array').value is Array
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Array').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Array').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Array').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Array').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ArrayBuffer').value is ArrayBuffer
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ArrayBuffer').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ArrayBuffer').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ArrayBuffer').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ArrayBuffer').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').value is Blob
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Boolean').value is Boolean
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Boolean').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Boolean').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Boolean').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Boolean').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ByteLengthQueuingStrategy').value is ByteLengthQueuingStrategy
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ByteLengthQueuingStrategy').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ByteLengthQueuingStrategy').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ByteLengthQueuingStrategy').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ByteLengthQueuingStrategy').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'CountQueuingStrategy').value is CountQueuingStrategy
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'CountQueuingStrategy').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'CountQueuingStrategy').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'CountQueuingStrategy').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'CountQueuingStrategy').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Crypto').value is Crypto
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Crypto').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Crypto').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Crypto').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Crypto').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'CryptoKey').value is CryptoKey
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'CryptoKey').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'CryptoKey').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'CryptoKey').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'CryptoKey').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMException').value is DOMException
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMException').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMException').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMException').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMException').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMPoint').value is DOMPoint
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMPoint').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMPoint').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMPoint').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMPoint').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMPointReadOnly').value is DOMPointReadOnly
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMPointReadOnly').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMPointReadOnly').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMPointReadOnly').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMPointReadOnly').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMRect').value is DOMRect
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMRect').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMRect').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMRect').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMRect').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMRectReadOnly').value is DOMRectReadOnly
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMRectReadOnly').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMRectReadOnly').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMRectReadOnly').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DOMRectReadOnly').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DataView').value is DataView
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DataView').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DataView').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DataView').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DataView').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Date').value is Date
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Date').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Date').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Date').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Date').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DedicatedWorkerGlobalScope').value is DedicatedWorkerGlobalScope
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DedicatedWorkerGlobalScope').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DedicatedWorkerGlobalScope').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DedicatedWorkerGlobalScope').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DedicatedWorkerGlobalScope').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Error').value is Error
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Error').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Error').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Error').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Error').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ErrorEvent').value is ErrorEvent
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ErrorEvent').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ErrorEvent').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ErrorEvent').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ErrorEvent').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EvalError').value is EvalError
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EvalError').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EvalError').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EvalError').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EvalError').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Event').value is Event
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Event').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Event').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Event').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Event').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').value is EventSource
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventTarget').value is EventTarget
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventTarget').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventTarget').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventTarget').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventTarget').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReader').value is FileReader
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReader').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReader').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReader').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReader').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReaderSync').value is FileReaderSync
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReaderSync').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReaderSync').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReaderSync').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReaderSync').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float32Array').value is Float32Array
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float32Array').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float32Array').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float32Array').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float32Array').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float64Array').value is Float64Array
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float64Array').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float64Array').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float64Array').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float64Array').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Function').value is Function
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Function').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Function').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Function').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Function').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Headers').value is Headers
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Headers').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Headers').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Headers').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Headers').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ImageData').value is ImageData
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ImageData').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ImageData').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ImageData').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ImageData').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int16Array').value is Int16Array
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int16Array').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int16Array').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int16Array').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int16Array').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int32Array').value is Int32Array
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int32Array').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int32Array').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int32Array').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int32Array').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int8Array').value is Int8Array
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int8Array').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int8Array').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int8Array').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int8Array').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Map').value is Map
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Map').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Map').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Map').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Map').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').value is MessageChannel
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').value is MessageEvent
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessagePort').value is MessagePort
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessagePort').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessagePort').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessagePort').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessagePort').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Number').value is Number
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Number').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Number').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Number').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Number').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Object').value is Object
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Object').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Object').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Object').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Object').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Performance').value is Performance
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Performance').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Performance').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Performance').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Performance').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceEntry').value is PerformanceEntry
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceEntry').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceEntry').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceEntry').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceEntry').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceMark').value is PerformanceMark
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceMark').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceMark').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceMark').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceMark').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceMeasure').value is PerformanceMeasure
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceMeasure').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceMeasure').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceMeasure').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceMeasure').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceObserver').value is PerformanceObserver
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceObserver').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceObserver').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceObserver').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceObserver').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceObserverEntryList').value is PerformanceObserverEntryList
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceObserverEntryList').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceObserverEntryList').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceObserverEntryList').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'PerformanceObserverEntryList').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ProgressEvent').value is ProgressEvent
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ProgressEvent').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ProgressEvent').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ProgressEvent').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ProgressEvent').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Promise').value is Promise
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Promise').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Promise').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Promise').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Promise').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'RangeError').value is RangeError
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'RangeError').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'RangeError').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'RangeError').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'RangeError').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ReadableStream').value is ReadableStream
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ReadableStream').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ReadableStream').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ReadableStream').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ReadableStream').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ReferenceError').value is ReferenceError
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ReferenceError').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ReferenceError').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ReferenceError').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ReferenceError').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'RegExp').value is RegExp
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'RegExp').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'RegExp').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'RegExp').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'RegExp').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Request').value is Request
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Request').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Request').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Request').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Request').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Response').value is Response
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Response').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Response').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Response').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Response').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Set').value is Set
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Set').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Set').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Set').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Set').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SharedArrayBuffer').value is SharedArrayBuffer
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SharedArrayBuffer').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SharedArrayBuffer').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SharedArrayBuffer').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SharedArrayBuffer').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'String').value is String
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'String').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'String').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'String').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'String').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SubtleCrypto').value is SubtleCrypto
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SubtleCrypto').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SubtleCrypto').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SubtleCrypto').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SubtleCrypto').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Symbol').value is Symbol
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Symbol').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Symbol').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Symbol').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Symbol').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SyntaxError').value is SyntaxError
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SyntaxError').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SyntaxError').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SyntaxError').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'SyntaxError').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'TextDecoder').value is TextDecoder
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'TextDecoder').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'TextDecoder').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'TextDecoder').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'TextDecoder').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'TextEncoder').value is TextEncoder
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'TextEncoder').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'TextEncoder').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'TextEncoder').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'TextEncoder').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'TypeError').value is TypeError
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'TypeError').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'TypeError').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'TypeError').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'TypeError').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URIError').value is URIError
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URIError').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URIError').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URIError').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URIError').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URL').value is URL
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URL').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URL').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URL').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URL').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URLSearchParams').value is URLSearchParams
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URLSearchParams').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URLSearchParams').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URLSearchParams').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URLSearchParams').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint16Array').value is Uint16Array
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint16Array').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint16Array').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint16Array').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint16Array').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint32Array').value is Uint32Array
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint32Array').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint32Array').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint32Array').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint32Array').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8Array').value is Uint8Array
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8Array').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8Array').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8Array').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8Array').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8ClampedArray').value is Uint8ClampedArray
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8ClampedArray').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8ClampedArray').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8ClampedArray').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8ClampedArray').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WeakMap').value is WeakMap
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WeakMap').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WeakMap').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WeakMap').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WeakMap').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WeakSet').value is WeakSet
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WeakSet').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WeakSet').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WeakSet').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WeakSet').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').value is WebSocket
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').value is WorkerGlobalScope
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').value is WorkerLocation
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerNavigator').value is WorkerNavigator
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerNavigator').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerNavigator').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerNavigator').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerNavigator').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').value is XMLHttpRequest
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequestEventTarget').value is XMLHttpRequestEventTarget
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequestEventTarget').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequestEventTarget').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequestEventTarget').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequestEventTarget').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitURL').value is webkitURL
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitURL').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitURL').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitURL').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitURL').configurable is true
PASS successfullyParsed is true

TEST COMPLETE
